Merge nil values by default
Previously, if you merged in {a: nil}, it would not overwrite an
existing value.  In #13, this was found to be an undesirable default
behavior.

Now, we will merge nil values by default. You can retain the old
behavior via `config.merge_nil_values = false` in your Config initializer.

This required changes to deep merge, found in: https://github.com/danielsdeleo/deep_merge/pull/33
This was released in deep_merge 1.2.0+.

Fixes #13