Add explicit implementation of Config::Options#as_json
Before, `Config::Options#as_json` was not explicitly defined, but the
class included the `Enumerable` module. So, if you were using
ActiveSupport Core Extensions, then the definition of `#as_json` for
Enumerable was invoked, which would represent the `Config::Options` as a
Array of key-value pairs instead of a Hash.

This changes brings `#as_json` in line with the output we get from
`to_json`: essentially, calling these methods on an instance of
`Config::Options` behaves the same as if they were called instead on
the hash returned by `to_hash`.