Raise explicit exception if two environment variables conflict
Suppose the user sets the variables `Settings.backend_database` in
addition to `Settings.backend_database.user`. Since it is possible to
override existing settings with env vars, one possible way to handle
this is to process the environment variables in a well-defined order and
allow lexicographically greater keys to override lower ones. However,
it's probably not what the user intended to happen, so instead we raise
a descriptive error when a conflict is detected.

This is likely not a breaking change because a `RuntimeError` is already
raised in this case, it's just not very friendly:

    TypeError:
      no implicit conversion of Symbol into Integer

Now we'll raise

    RuntimeError:
        Environment variable Settings.backend_database.user conflicts
        with variable Settings.backend_database