Support arbitrary sources in Config.load_files and Config.load_and_set_settings (#315)
* Add instructions for running specs

* Backfile test for Pathname support in load_files

There's a .to_s call in there that's likely to support these kinds of
objects, but there wasn't a test for it.

* Rename `file` variables to `source`

In an upcoming change, we'll allow arbirtary sources as parameters to
these methods, instead of just filenames. The method names remain
unchanged for now for backwards compatibility.

* Support arbitrary Source objects in Config.load_files

* Add test for loading HashSource from load_files

* Remove uniq call in load_files

This uniq call causes issues when mixing the types of sources. This
change can technically break some applications that were depending on
the de-deduplication of source files.

* Remove .to_s call in load_files on source

This change effectively conslidates the logic for coercing source
specifications to Source objects into `Config::Options#add_source!`.
For sources specified as Strings or Pathnames, there is effectively no
change, but for sources specified as something else that responds to
`.to_s`, this change could be breaking. In those cases, application
developers can explicitly call `.to_s` on those objects before handing
them to config.

* Add CHANGELOG entry